refactor: Adding ViewSetting model
